backend/requirements.txt
# FastAPI
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Validation
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Testing
pytest==7.4.4
pytest-asyncio==0.23.4
httpx==0.26.0